requirements.txt
ultralytics~=8.3.23
torch>=1.7.0        # PyTorch, para suportar operações de deep learning
opencv-python>=4.1.1 # OpenCV para processamento de imagens e vídeos
numpy>=1.18.5       # Operações matemáticas e de matriz
matplotlib>=3.2.2   # Para visualização de resultados e gráficos
pillow>=7.1.2       # Biblioteca de processamento de imagens
tqdm>=4.64.0        # Barra de progresso
PyYAML>=5.3.1       # Para ler arquivos de configuração no formato YAML
requests>=2.23.0    # Para fazer requisições HTTP, como baixar modelos pré-treinados
seaborn>=0.11.0     # Para gráficos e visualização de dados

ipython~=8.29.0
GitPython~=3.1.43
gitdb~=4.0.11
pip~=23.2.1
protobuf~=5.28.3
filelock~=3.16.1
absl-py~=2.1.0
grpcio~=1.67.0
parso~=0.8.4
jedi~=0.19.1
pika~=1.3.2
pytz~=2024.2
thop~=0.1.1-2209072238
ultralytics-thop~=2.0.9
smmap~=5.0.1
omegaconf~=2.3.0
hydra-core~=1.3.2
antlr4-python3-runtime~=4.9.3
typing_extensions~=4.12.2
packaging~=24.1
charset-normalizer~=3.4.0
setuptools~=75.2.0
scipy~=1.14.1
mpmath~=1.3.0
sympy~=1.13.1
decorator~=5.1.1
tensorboard~=2.18.0
triton~=3.1.0
torchvision~=0.20.0
psutil~=6.1.0
urllib3~=2.2.3
fsspec~=2024.10.0
MarkupSafe~=3.0.2
pandas~=2.2.3
python-dateutil~=2.9.0.post0
Pygments~=2.18.0
traitlets~=5.14.3
stack-data~=0.6.3
matplotlib-inline~=0.1.7
prompt_toolkit~=3.0.48
wcwidth~=0.2.13
certifi~=2024.8.30
pexpect~=4.9.0
ptyprocess~=0.7.0
cycler~=0.12.1
six~=1.16.0
Markdown~=3.7
networkx~=3.4.2
idna~=3.10
asttokens~=2.4.1
contourpy~=1.3.0
executing~=2.1.0
fonttools~=4.54.1
pure_eval~=0.2.3
pyparsing~=3.2.0
Jinja2~=3.1.4
kiwisolver~=1.4.7
Werkzeug~=3.0.6
torchaudio~=2.5.0+cu124
qrcode~=8.0
pyzbar~=0.1.9
segno~=1.6.1
docker~=7.1.0
Flask~=3.1.0